+++ lib_paysages/tools.h
@@ -8,11 +8,11 @@
 extern "C" {
 #endif
 
-double toolsRandom();
-double toolsBicubicInterpolate(double stencil[16], double x, double y);
-void toolsFloat2DMapCopy(double* src, double* dest, int src_xstart, int src_ystart, int dest_xstart, int dest_ystart, int xsize, int ysize, int src_xstep, int src_ystep, int dest_xstep, int dest_ystep);
+float toolsRandom();
+float toolsBicubicInterpolate(float stencil[16], float x, float y);
+void toolsFloat2DMapCopy(float* src, float* dest, int src_xstart, int src_ystart, int dest_xstart, int dest_ystart, int xsize, int ysize, int src_xstep, int src_ystep, int dest_xstep, int dest_ystep);
 Vector3 toolsGetNormalFromTriangle(Vector3 center, Vector3 bottom, Vector3 right);
-double toolsGetDistance2D(double x1, double y1, double x2, double y2);
+float toolsGetDistance2D(float x1, float y1, float x2, float y2);
 void materialSave(PackStream* stream, SurfaceMaterial* material);
 void materialLoad(PackStream* stream, SurfaceMaterial* material);
